+++ CHANGELOG.rst
@@ -7,6 +7,8 @@ master
 - Read in daily Spectronus data and added to plots
 - Added subset command for extracting a data for a single instrument
 - Remove nan values from met data
+- Added backup sync command for syncronising the backups with the external drives
+    which are removed from the vessel after the voyage
 
 0.1.1
 -----